mypy fixes, refactor filewriters,
