+++ airbyte-lib/airbyte_lib/caches/__init__.py
@@ -1,12 +1,13 @@
 """Base module for all caches."""
 
-from airbyte_lib.caches.parquet import ParquetWriter
 from airbyte_lib.caches.sql import SQLCache
 from airbyte_lib.caches.sql.types import SQLTypeConverter
+from airbyte_lib.file_writers.parquet import ParquetWriter, ParquetWriterConfig
 
 # We export these classes for easy access: `airbyte_lib.caches...`
 __all__ = [
     "ParquetWriter",
+    "ParquetWriterConfig",
     "SQLCache",
     "SQLTypeConverter",
 ]
