+++ airbyte-lib/airbyte_lib/caches/bases/__init__.py
@@ -1,11 +1,9 @@
 from .config import CacheConfigBase
-from .core import BaseCache
-from .file_writers import FileWriterBase
+from .core import CacheBase
 from .sql import SQLCache, SQLCacheConfigBase
 
 __all__ = [
-    "BaseCache",
-    "FileWriterBase",
+    "CacheBase",
     "CacheConfigBase",
     "SQLCache",
     "SQLCacheConfigBase",
