+++ airbyte-lib/airbyte_lib/file_writers/__init__.py
@@ -1,0 +1,8 @@
+from .base import FileWriterBase
+from .parquet import ParquetWriter, ParquetWriterConfig
+
+__all__ = [
+    "FileWriterBase",
+    "ParquetWriter", 
+    "ParquetWriterConfig",
+]
